// src/utils/enums/dataExchangeStatusEnum.ts
export enum DataExchangeStatusEnum {
    EXPORT_SUCCESS = 'EXPORT_SUCCESS',
    IMPORT_SUCCESS = 'IMPORT_SUCCESS',
    PROVIDER_EXPORT_ERROR = 'PROVIDER_EXPORT_ERROR',
    CONSENT_IMPORT_ERROR = 'CONSENT_IMPORT_ERROR',
    CONSENT_EXPORT_ERROR = 'CONSENT_EXPORT_ERROR',
    CONSUMER_IMPORT_ERROR = 'CONSUMER_IMPORT_ERROR',
    UNDEFINED_ERROR = 'UNDEFINED_ERROR',
    PEP_ERROR = 'PEP_ERROR',
    NODE_CALLBACK_ERROR = 'NODE_CALLBACK_ERROR',
    PENDING = 'PENDING',
    TRANSFER_STARTED = 'TRANSFER_STARTED',
    TRANSFER_COMPLETED = 'TRANSFER_COMPLETED',
    TRANSFER_FAILED = 'TRANSFER_FAILED',
}